Ensure event handlers are triggered, not DOM functions
Chrome 36 introduces a change that adds an event.animate() function to the DOM. The jQuery trigger() function will execute functions and not event handlers where a naming clash exists.

Changing the code to use the jQuery triggerHandler() function rather than trigger() ensures that event handlers are always fired as intended.